Broke global Actions into small  actions for each component
