+++ src/redux/Actions/workActions.js
@@ -1,0 +1,45 @@
+export const workNameChange = value => ({
+    type: "CHANGE_WORK_NAME",
+    payload: value
+  })
+  
+  export const workDescriptionChange = value => ({
+    type: "CHANGE_WORK_DESCRIPTION",
+    payload: value
+  })
+  
+  export const workPositionChange = value => ({
+    type: "CHANGE_WORK_POSITION",
+    payload: value
+  
+  })
+  
+  export const workEarlyYearChange = value => ({
+    type: "CHANGE_WORK_EARLY",
+    payload: value
+  })
+  
+  export const workLateYearChange = value => ({
+    type: "CHANGE_WORK_LATE",
+    payload: value
+  })
+  
+  export const workCountryChange = value => ({
+    type: "CHANGE_WORK_COUNTRY",
+    payload: value
+  })
+  
+  export const workCityChange = value => ({
+    type: "CHANGE_WORK_CITY",
+    payload: value
+  
+  })
+  
+  export const workRemove = value => ({
+    type: "REMOVE_WORK",
+  })
+  
+  export const workAdd = value => ({
+    type: "ADD_WORK",
+  })
+
