+++ src/redux/Reducers/certificateDataReducer.js
@@ -1,13 +1,23 @@
-const certificateDataReducer = (state = [{certificate:""}], action) => {
+const initialState = {
+    inputValue: '',
+    certificates: []
+};
+const certificateDataReducer = (state = initialState, action) => {
     switch (action.type) {
+        case 'SET_CERT_INPUT':
+            return {
+                ...state,
+                inputValue: action.payload
+            }
+
         case "CHANGE_CERT": {
-            return [{...state , certificate: action.payload }]
+            return { ...state, certificate: action.payload }
         }
         case "ADD_CERT":
-            {
-                return [...state,
-                { certificate: action.payload }
-                ]
+            return {
+                ...state,
+                inputValue: '',
+                certificates: [...state.certificates, state.inputValue]
             }
         case "REMOVE_CERT": {
             return state.slice(action.index).concat(
